+++ packages/jupytercad-extension/src/index.ts
@@ -1,143 +1,150 @@
-import {
-  ILayoutRestorer,
-  JupyterFrontEnd,
-  JupyterFrontEndPlugin
-} from '@jupyterlab/application';
-import { WidgetTracker } from '@jupyterlab/apputils';
-import { ITranslator } from '@jupyterlab/translation';
-import { IMainMenu } from '@jupyterlab/mainmenu';
+// import {
+//   ILayoutRestorer,
+//   JupyterFrontEnd,
+//   JupyterFrontEndPlugin
+// } from '@jupyterlab/application';
+// import { WidgetTracker } from '@jupyterlab/apputils';
+// import { ITranslator } from '@jupyterlab/translation';
+// import { IMainMenu } from '@jupyterlab/mainmenu';
 
-import fcplugin from './fcplugin/plugins';
-import jcadPlugin from './jcadplugin/plugins';
-import { JupyterCadModel } from './model';
-import { ControlPanelModel } from './panelview/model';
-import { LeftPanelWidget } from './panelview/leftpanel';
-import { RightPanelWidget } from './panelview/rightpanel';
-import {
-  IJupyterCadDocTracker,
-  IJupyterCadTracker,
-  IAnnotation
-} from './token';
-import { jcLightIcon } from './tools';
-import { IAnnotationModel } from './types';
-import { JupyterCadWidget } from './widget';
-import { AnnotationModel } from './annotation/model';
-import { yJupyterCADWidgetPlugin } from './notebookrenderer';
-import { addCommands, CommandIDs } from './commands';
+// import fcplugin from './fcplugin/plugins';
+// import jcadPlugin from './jcadplugin/plugins';
+// import { JupyterCadModel } from './model';
+// import { ControlPanelModel } from './panelview/model';
+// import { LeftPanelWidget } from './panelview/leftpanel';
+// import { RightPanelWidget } from './panelview/rightpanel';
+// import {
+//   IJupyterCadDocTracker,
+//   IJupyterCadTracker,
+//   IAnnotation
+// } from './token';
+// import { jcLightIcon } from './tools';
+// import { IAnnotationModel } from './types';
+// import { JupyterCadWidget } from './widget';
+// import { AnnotationModel } from './annotation/model';
+// import { yJupyterCADWidgetPlugin } from './notebookrenderer';
+// import { addCommands, CommandIDs } from './commands';
 
-const NAME_SPACE = 'jupytercad';
+export * from './token';
+export * from './types';
+export * from './model';
+export * from './widget';
+export * from './annotation';
 
-const plugin: JupyterFrontEndPlugin<IJupyterCadTracker> = {
-  id: 'jupytercad:plugin',
-  autoStart: true,
-  requires: [ITranslator],
-  optional: [IMainMenu],
-  provides: IJupyterCadDocTracker,
-  activate: (
-    app: JupyterFrontEnd,
-    translator: ITranslator,
-    mainMenu?: IMainMenu
-  ): IJupyterCadTracker => {
-    const tracker = new WidgetTracker<JupyterCadWidget>({
-      namespace: NAME_SPACE
-    });
-    JupyterCadModel.worker = new Worker(
-      new URL('./worker', (import.meta as any).url)
-    );
+// const plugin: JupyterFrontEndPlugin<IJupyterCadTracker> = {
+//   id: 'jupytercad:plugin',
+//   autoStart: true,
+//   requires: [ITranslator],
+//   optional: [IMainMenu],
+//   provides: IJupyterCadDocTracker,
+//   activate: (
+//     app: JupyterFrontEnd,
+//     translator: ITranslator,
+//     mainMenu?: IMainMenu
+//   ): IJupyterCadTracker => {
+//     const tracker = new WidgetTracker<JupyterCadWidget>({
+//       namespace: NAME_SPACE
+//     });
+//     JupyterCadModel.worker = new Worker(
+//       new URL('./worker', (import.meta as any).url)
+//     );
 
-    console.log('JupyterLab extension jupytercad is activated!');
+//     console.log('JupyterLab extension jupytercad is activated!');
 
-    /**
-     * Whether there is an active notebook.
-     */
-    const isEnabled = (): boolean => {
-      return (
-        tracker.currentWidget !== null &&
-        tracker.currentWidget === app.shell.currentWidget
-      );
-    };
+//     /**
+//      * Whether there is an active notebook.
+//      */
+//     const isEnabled = (): boolean => {
+//       return (
+//         tracker.currentWidget !== null &&
+//         tracker.currentWidget === app.shell.currentWidget
+//       );
+//     };
 
-    addCommands(app, tracker, translator);
-    if (mainMenu) {
-      populateMenus(mainMenu, isEnabled);
-    }
+//     addCommands(app, tracker, translator);
+//     if (mainMenu) {
+//       populateMenus(mainMenu, isEnabled);
+//     }
 
-    return tracker;
-  }
-};
+//     return tracker;
+//   }
+// };
 
-const annotationPlugin: JupyterFrontEndPlugin<IAnnotationModel> = {
-  id: 'jupytercad:annotation',
-  autoStart: true,
-  requires: [IJupyterCadDocTracker],
-  provides: IAnnotation,
-  activate: (app: JupyterFrontEnd, tracker: IJupyterCadTracker) => {
-    const annotationModel = new AnnotationModel({
-      context: tracker.currentWidget?.context
-    });
+// const annotationPlugin: JupyterFrontEndPlugin<IAnnotationModel> = {
+//   id: 'jupytercad:annotation',
+//   autoStart: true,
+//   requires: [IJupyterCadDocTracker],
+//   provides: IAnnotation,
+//   activate: (app: JupyterFrontEnd, tracker: IJupyterCadTracker) => {
+//     const annotationModel = new AnnotationModel({
+//       context: tracker.currentWidget?.context
+//     });
 
-    tracker.currentChanged.connect((_, changed) => {
-      annotationModel.context = changed?.context || undefined;
-    });
-    return annotationModel;
-  }
-};
-
-const controlPanel: JupyterFrontEndPlugin<void> = {
-  id: 'jupytercad:controlpanel',
-  autoStart: true,
-  requires: [ILayoutRestorer, IJupyterCadDocTracker, IAnnotation],
-  activate: (
-    app: JupyterFrontEnd,
-    restorer: ILayoutRestorer,
-    tracker: IJupyterCadTracker,
-    annotationModel: IAnnotationModel
-  ) => {
-    const controlModel = new ControlPanelModel({ tracker });
-
-    const leftControlPanel = new LeftPanelWidget({
-      model: controlModel,
-      annotationModel,
-      tracker
-    });
-    leftControlPanel.id = 'jupytercad::leftControlPanel';
-    leftControlPanel.title.caption = 'JupyterCad Control Panel';
-    leftControlPanel.title.icon = jcLightIcon;
-
-    const rightControlPanel = new RightPanelWidget({ model: controlModel });
-    rightControlPanel.id = 'jupytercad::rightControlPanel';
-    rightControlPanel.title.caption = 'JupyterCad Control Panel';
-    rightControlPanel.title.icon = jcLightIcon;
-
-    if (restorer) {
-      restorer.add(leftControlPanel, NAME_SPACE);
-      restorer.add(rightControlPanel, NAME_SPACE);
-    }
-    app.shell.add(leftControlPanel, 'left', { rank: 2000 });
-    app.shell.add(rightControlPanel, 'right', { rank: 2000 });
-  }
-};
+//     tracker.currentChanged.connect((_, changed) => {
+//       annotationModel.context = changed?.context || undefined;
+//     });
+//     return annotationModel;
+//   }
+// };
 
 /**
- * Populates the application menus for the notebook.
+ * @TODO Move control panel to jupytercad-lab package
  */
-function populateMenus(mainMenu: IMainMenu, isEnabled: () => boolean): void {
-  // Add undo/redo hooks to the edit menu.
-  mainMenu.editMenu.undoers.redo.add({
-    id: CommandIDs.redo,
-    isEnabled
-  });
-  mainMenu.editMenu.undoers.undo.add({
-    id: CommandIDs.undo,
-    isEnabled
-  });
-}
+// const controlPanel: JupyterFrontEndPlugin<void> = {
+//   id: 'jupytercad:controlpanel',
+//   autoStart: true,
+//   requires: [ILayoutRestorer, IJupyterCadDocTracker, IAnnotation],
+//   activate: (
+//     app: JupyterFrontEnd,
+//     restorer: ILayoutRestorer,
+//     tracker: IJupyterCadTracker,
+//     annotationModel: IAnnotationModel
+//   ) => {
+//     const controlModel = new ControlPanelModel({ tracker });
 
-export default [
-  plugin,
-  controlPanel,
-  fcplugin,
-  jcadPlugin,
-  annotationPlugin,
-  yJupyterCADWidgetPlugin
-];
+//     const leftControlPanel = new LeftPanelWidget({
+//       model: controlModel,
+//       annotationModel,
+//       tracker
+//     });
+//     leftControlPanel.id = 'jupytercad::leftControlPanel';
+//     leftControlPanel.title.caption = 'JupyterCad Control Panel';
+//     leftControlPanel.title.icon = jcLightIcon;
+
+//     const rightControlPanel = new RightPanelWidget({ model: controlModel });
+//     rightControlPanel.id = 'jupytercad::rightControlPanel';
+//     rightControlPanel.title.caption = 'JupyterCad Control Panel';
+//     rightControlPanel.title.icon = jcLightIcon;
+
+//     if (restorer) {
+//       restorer.add(leftControlPanel, NAME_SPACE);
+//       restorer.add(rightControlPanel, NAME_SPACE);
+//     }
+//     app.shell.add(leftControlPanel, 'left', { rank: 2000 });
+//     app.shell.add(rightControlPanel, 'right', { rank: 2000 });
+//   }
+// };
+
+// /**
+//  * Populates the application menus for the notebook.
+//  */
+// function populateMenus(mainMenu: IMainMenu, isEnabled: () => boolean): void {
+//   // Add undo/redo hooks to the edit menu.
+//   mainMenu.editMenu.undoers.redo.add({
+//     id: CommandIDs.redo,
+//     isEnabled
+//   });
+//   mainMenu.editMenu.undoers.undo.add({
+//     id: CommandIDs.undo,
+//     isEnabled
+//   });
+// }
+
+// export default [
+//   plugin,
+//   controlPanel,
+//   fcplugin,
+//   jcadPlugin,
+//   annotationPlugin,
+//   yJupyterCADWidgetPlugin
+// ];
