+++ python/jupytercad-core/package.json
@@ -48,9 +48,12 @@
         "stylelint:check": "stylelint --cache \"style/**/*.css\"",
         "watch": "run-p watch:src watch:labextension",
         "watch:src": "tsc -w --sourceMap",
-        "watch:labextension": "jupyter labextension watch ."
+        "watch:labextension": "jupyter labextension watch .",
+        "build:worker": "webpack --config worker.webpack.config.js --mode=development",
+        "build:worker:prod": "webpack --config worker.webpack.config.js --mode=production"
     },
     "dependencies": {
+        "@jupytercad/jupytercad-extension": "^0.3.3",
         "@jupyterlab/application": "^4.0.0",
         "@jupyterlab/coreutils": "^6.0.0",
         "@jupyterlab/services": "^7.0.0"
