+++ python/jupytercad-core/pyproject.toml
@@ -23,7 +23,9 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
 ]
 dependencies = [
-    "jupyter_server>=2.0.1,<3"
+    "jupyter_server>=2.0.6,<3",
+    "jupyter_ydoc>=1.1.0,<2",
+    "y-py>=0.6.0,<0.7"
 ]
 dynamic = ["version", "description", "authors", "urls", "keywords"]
